+++ packages/url_launcher/url_launcher/lib/url_launcher.dart
@@ -117,8 +117,10 @@ Future<bool> launch(
 ///
 /// On Android (from API 30), [canLaunch] will return `false` when the required
 /// visibility configuration is not provided in the AndroidManifest.xml file.
-/// For more information see the [Managing package visibility](https://developer.android.com/training/basics/intents/package-visibility)
-/// article in the Android docs.
+/// For more information see the
+/// [Package visibility filtering on Android](https://developer.android.com/training/basics/intents/package-visibility)
+/// article in the Android documentation or the url_launcher example app's
+/// [AndroidManifest.xml's queries element](https://github.com/flutter/plugins/blob/master/packages/url_launcher/url_launcher/example/android/app/src/main/AndroidManifest.xml).
 Future<bool> canLaunch(String urlString) async {
   return await UrlLauncherPlatform.instance.canLaunch(urlString);
 }